Crypto.Cipher.TripleDES

Description

Documentation

data DES_EEE3

3DES with 3 different keys used all in the same direction

data DES_EDE3

3DES with 3 different keys used in alternative direction

data DES_EEE2

3DES where the first and third keys are equal, used in the same direction

data DES_EDE2

3DES where the first and third keys are equal, used in alternative direction