Crypto.Hash.IO

Description

Generalized impure cryptographic hash interface

Documentation

class HashAlgorithm a where

Class representing hashing algorithms.

The interface presented here is update in place and lowlevel. the Hash module takes care of hidding the mutable interface properly.

Associated Types

type HashBlockSize a :: Nat

Associated type for the block size of the hash algorithm

type HashDigestSize a :: Nat

Associated type for the digest size of the hash algorithm

type HashInternalContextSize a :: Nat

Associated type for the internal context size of the hash algorithm

Methods

hashBlockSize :: a -> Int

Get the block size of a hash algorithm

hashDigestSize :: a -> Int

Get the digest size of a hash algorithm

hashInternalContextSize :: a -> Int

Get the size of the context used for a hash algorithm

hashInternalInit :: Ptr (Context a) -> IO ()

Initialize a context pointer to the initial state of a hash algorithm

hashInternalUpdate :: Ptr (Context a) -> Ptr Word8 -> Word32 -> IO ()

Update the context with some raw data

hashInternalFinalize :: Ptr (Context a) -> Ptr (Digest a) -> IO ()

Finalize the context and set the digest raw memory to the right value

data MutableContext a

A Mutable hash context

This type is an instance of ByteArrayAccess for debugging purpose. Internal layout is architecture dependent, may contain uninitialized data fragments, and change in future versions. The bytearray should not be used as input to cryptographic algorithms.

hashMutableInit :: HashAlgorithm alg => IO (MutableContext alg)

Create a new mutable hash context.

the algorithm used is automatically determined from the return constraint.

hashMutableInitWith :: HashAlgorithm alg => alg -> IO (MutableContext alg)

Create a new mutable hash context.

The algorithm is explicitely passed as parameter

hashMutableUpdate :: (ByteArrayAccess ba, HashAlgorithm a) => MutableContext a -> ba -> IO ()

Update a mutable hash context in place

hashMutableFinalize :: forall a. HashAlgorithm a => MutableContext a -> IO (Digest a)

Finalize a mutable hash context and compute a digest

hashMutableReset :: HashAlgorithm a => MutableContext a -> IO ()

Reset the mutable context to the initial state of the hash