Crypto.KDF.BCryptPBKDF

Description

Port of the bcrypt_pbkdf key derivation function from OpenBSD as described at http://man.openbsd.org/bcrypt_pbkdf.3.

Documentation

data Parameters

Constructors

Parameters

Fields

iterCounts :: Int

The number of user-defined iterations for the algorithm (must be > 0)

outputLength :: Int

The number of bytes to generate out of BCryptPBKDF (must be in 1..1024)

generate :: (ByteArray pass, ByteArray salt, ByteArray output) => Parameters -> pass -> salt -> output

Derive a key of specified length using the bcrypt_pbkdf algorithm.

hashInternal :: (ByteArrayAccess pass, ByteArrayAccess salt, ByteArray output) => pass -> salt -> output

Internal hash function used by generate.

Normal users should not need this.