Crypto.MAC.KeyedBlake2

Description

Expose a MAC interface to the keyed Blake2 algorithms defined in RFC 7693.

Documentation

class HashAlgorithm a => HashBlake2 a

Typeclass for the Blake2 family of digest functions.

Minimal complete definition

blake2InternalKeyedInit

newtype KeyedBlake2 a

Represent a Blake2b MAC that is a phantom type with the hash used to produce the MAC.

The Eq instance is constant time. No Show instance is provided, to avoid printing by mistake.

Constructors

KeyedBlake2

Fields

keyedBlake2GetDigest :: Digest a

keyedBlake2 :: (HashBlake2 a, ByteArrayAccess key, ByteArrayAccess ba) => key -> ba -> KeyedBlake2 a

Compute a Blake2 MAC using the supplied key.

keyedBlake2Lazy :: (HashBlake2 a, ByteArrayAccess key) => key -> ByteString -> KeyedBlake2 a

Compute a Blake2 MAC using the supplied key, for a lazy input.

Incremental

data Context a

Represent an ongoing Blake2 state, that can be appended with update and finalized to a KeyedBlake2 with finalize.

initialize :: forall a key. (HashBlake2 a, ByteArrayAccess key) => key -> Context a

Initialize a new incremental keyed Blake2 context with the supplied key.

update :: (HashBlake2 a, ByteArrayAccess ba) => Context a -> ba -> Context a

Incrementally update a keyed Blake2 context.

updates :: (HashBlake2 a, ByteArrayAccess ba) => Context a -> [ba] -> Context a

Incrementally update a keyed Blake2 context with multiple inputs.

finalize :: HashBlake2 a => Context a -> KeyedBlake2 a

Finalize a keyed Blake2 context and return the computed MAC.