Crypto.MAC.Poly1305

Description

Poly1305 implementation

Documentation

type Ctx = State

Deprecated: use Poly1305 State instead

Poly1305 State. use State instead of Ctx

data State

Poly1305 State

This type is an instance of ByteArrayAccess for debugging purpose. Internal layout is architecture dependent, may contain uninitialized data fragments, and change in future versions. The bytearray should not be used as input to cryptographic algorithms.

newtype Auth

Poly1305 Auth

Constructors

Auth Bytes

authTag :: ByteArrayAccess b => b -> CryptoFailable Auth

Incremental MAC Functions

initialize :: ByteArrayAccess key => key -> CryptoFailable State

initialize a Poly1305 context

update :: ByteArrayAccess ba => State -> ba -> State

update a context with a bytestring

updates :: ByteArrayAccess ba => State -> [ba] -> State

updates a context with multiples bytestring

finalize :: State -> Auth

finalize the context into a digest bytestring

One-pass MAC function

auth :: (ByteArrayAccess key, ByteArrayAccess ba) => key -> ba -> Auth

One-pass authorization creation