Crypto.Number.Generate

Description

Documentation

data GenTopPolicy

Top bits policy when generating a number

Constructors

SetHighest
set the highest bit

SetTwoHighest
set the two highest bit

generateParams

Arguments

:: MonadRandom m

=> Int
number of bits

-> Maybe GenTopPolicy
top bit policy

-> Bool
force the number to be odd

-> m Integer

Generate a number for a specific size of bits, and optionaly set bottom and top bits

If the top bit policy is Nothing, then nothing is done on the highest bit (it's whatever the random generator set).

If @generateOdd is set to True, then the number generated is guaranteed to be odd. Otherwise it will be whatever is generated

generateMax

Arguments

:: MonadRandom m

=> Integer
range

-> m Integer

Generate a positive integer x, s.t. 0 <= x < range

generateBetween :: MonadRandom m => Integer -> Integer -> m Integer

generate a number between the inclusive bound [low,high].