Crypto.PubKey.Curve448

Description

Curve448 support

Internally uses Decaf point compression to omit the cofactor and implementation by Mike Hamburg. Externally API and data types are compatible with the encoding specified in RFC 7748.

Documentation

data SecretKey

A Curve448 Secret key

data PublicKey

A Curve448 public key

data DhSecret

A Curve448 Diffie Hellman secret related to a public key and a secret key.

Smart constructors

dhSecret :: ByteArrayAccess b => b -> CryptoFailable DhSecret

Create a DhSecret from a bytearray object

publicKey :: ByteArrayAccess bs => bs -> CryptoFailable PublicKey

Try to build a public key from a bytearray

secretKey :: ByteArrayAccess bs => bs -> CryptoFailable SecretKey

Try to build a secret key from a bytearray

Methods

dh :: PublicKey -> SecretKey -> DhSecret

Compute the Diffie Hellman secret from a public key and a secret key.

This implementation may return an all-zero value as it does not check for the condition.

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key

generateSecretKey :: MonadRandom m => m SecretKey

Generate a secret key.