Crypto.PubKey.DSA

Description

An implementation of the Digital Signature Algorithm (DSA)

Documentation

data Params

Represent DSA parameters namely P, G, and Q.

Constructors

Params

Fields

params_p :: Integer

DSA p

params_g :: Integer

DSA g

params_q :: Integer

DSA q

data Signature

Represent a DSA signature namely R and S.

Constructors

Signature

Fields

sign_r :: Integer

DSA r

sign_s :: Integer

DSA s

data PublicKey

Represent a DSA public key.

Constructors

PublicKey

Fields

public_params :: Params

DSA parameters

public_y :: PublicNumber

DSA public Y

data PrivateKey

Represent a DSA private key.

Only x need to be secret. the DSA parameters are publicly shared with the other side.

Constructors

PrivateKey

Fields

private_params :: Params

DSA parameters

private_x :: PrivateNumber

DSA private X

type PublicNumber = Integer

DSA Public Number, usually embedded in DSA Public Key

type PrivateNumber = Integer

DSA Private Number, usually embedded in DSA Private Key

Generation

generatePrivate :: MonadRandom m => Params -> m PrivateNumber

generate a private number with no specific property this number is usually called X in DSA text.

calculatePublic :: Params -> PrivateNumber -> PublicNumber

Calculate the public number from the parameters and the private key

Signature primitive

sign :: (ByteArrayAccess msg, HashAlgorithm hash, MonadRandom m) => PrivateKey -> hash -> msg -> m Signature

sign message using the private key.

signWith

Arguments

:: (ByteArrayAccess msg, HashAlgorithm hash)

=> Integer
k random number

-> PrivateKey
private key

-> hash
hash function

-> msg
message to sign

-> Maybe Signature

sign message using the private key and an explicit k number.

Verification primitive

verify :: (ByteArrayAccess msg, HashAlgorithm hash) => hash -> PublicKey -> Signature -> msg -> Bool

verify a bytestring using the public key.

Key pair

data KeyPair

Represent a DSA key pair

Constructors

KeyPair Params PublicNumber PrivateNumber

toPublicKey :: KeyPair -> PublicKey

Public key of a DSA Key pair

toPrivateKey :: KeyPair -> PrivateKey

Private key of a DSA Key pair