Crypto.PubKey.ECC.ECDSA

Description

WARNING: Signature operations may leak the private key. Signature verification should be safe.

Documentation

data Signature

Represent a ECDSA signature namely R and S.

Constructors

Signature

Fields

sign_r :: Integer

ECDSA r

sign_s :: Integer

ECDSA s

type PublicPoint = Point

ECC Public Point

data PublicKey

ECDSA Public Key.

Constructors

PublicKey

Fields

public_curve :: Curve

public_q :: PublicPoint

type PrivateNumber = Integer

ECC Private Number

data PrivateKey

ECDSA Private Key.

Constructors

PrivateKey

Fields

private_curve :: Curve

private_d :: PrivateNumber

data KeyPair

ECDSA Key Pair.

Constructors

KeyPair Curve PublicPoint PrivateNumber

toPublicKey :: KeyPair -> PublicKey

Public key of a ECDSA Key pair.

toPrivateKey :: KeyPair -> PrivateKey

Private key of a ECDSA Key pair.

signWith

Arguments

:: (ByteArrayAccess msg, HashAlgorithm hash)

=> Integer
k random number

-> PrivateKey
private key

-> hash
hash function

-> msg
message to sign

-> Maybe Signature

Sign message using the private key and an explicit k number.

WARNING: Vulnerable to timing attacks.

signDigestWith

Arguments

:: HashAlgorithm hash

=> Integer
k random number

-> PrivateKey
private key

-> Digest hash
digest to sign

-> Maybe Signature

Sign digest using the private key and an explicit k number.

WARNING: Vulnerable to timing attacks.

sign :: (ByteArrayAccess msg, HashAlgorithm hash, MonadRandom m) => PrivateKey -> hash -> msg -> m Signature

Sign message using the private key.

WARNING: Vulnerable to timing attacks.

signDigest :: (HashAlgorithm hash, MonadRandom m) => PrivateKey -> Digest hash -> m Signature

Sign digest using the private key.

WARNING: Vulnerable to timing attacks.

verify :: (ByteArrayAccess msg, HashAlgorithm hash) => hash -> PublicKey -> Signature -> msg -> Bool

Verify a bytestring using the public key.

verifyDigest :: HashAlgorithm hash => PublicKey -> Signature -> Digest hash -> Bool

Verify a digest using the public key.