Crypto.PubKey.Ed448

Description

Ed448 support

Internally uses Decaf point compression to omit the cofactor and implementation by Mike Hamburg. Externally API and data types are compatible with the encoding specified in RFC 8032.

Documentation

data SecretKey

An Ed448 Secret key

data PublicKey

An Ed448 public key

data Signature

An Ed448 signature

Size constants

publicKeySize :: Int

A public key is 57 bytes

secretKeySize :: Int

A secret key is 57 bytes

signatureSize :: Int

A signature is 114 bytes

Smart constructors

signature :: ByteArrayAccess ba => ba -> CryptoFailable Signature

Try to build a signature from a bytearray

publicKey :: ByteArrayAccess ba => ba -> CryptoFailable PublicKey

Try to build a public key from a bytearray

secretKey :: ByteArrayAccess ba => ba -> CryptoFailable SecretKey

Try to build a secret key from a bytearray

Methods

toPublic :: SecretKey -> PublicKey

Create a public key from a secret key

sign :: ByteArrayAccess ba => SecretKey -> PublicKey -> ba -> Signature

Sign a message using the key pair

verify :: ByteArrayAccess ba => PublicKey -> ba -> Signature -> Bool

Verify a message

generateSecretKey :: MonadRandom m => m SecretKey

Generate a secret key