Crypto.PubKey.Rabin.Basic

Description

Rabin cryptosystem for public-key cryptography and digital signature.

Documentation

data PublicKey

Represent a Rabin public key.

Constructors

PublicKey

Fields

public_size :: Int

size of key in bytes

public_n :: Integer

public p*q

data PrivateKey

Represent a Rabin private key.

Constructors

PrivateKey

Fields

private_pub :: PublicKey

private_p :: Integer

p prime number

private_q :: Integer

q prime number

private_a :: Integer

private_b :: Integer

data Signature

Rabin Signature.

Constructors

Signature (Integer, Integer)

generate :: MonadRandom m => Int -> m (PublicKey, PrivateKey)

Generate a pair of (private, public) key of size in bytes. Primes p and q are both congruent 3 mod 4.

See algorithm 8.11 in "Handbook of Applied Cryptography" by Alfred J. Menezes et al.

encrypt

Arguments

:: (HashAlgorithm hash, MonadRandom m)

=> OAEPParams hash ByteString ByteString
OAEP padding parameters

-> PublicKey
public key

-> ByteString
plaintext

-> m (Either Error ByteString)

Encrypt plaintext using public key.

encryptWithSeed

Arguments

:: HashAlgorithm hash

=> ByteString
Seed

-> OAEPParams hash ByteString ByteString
OAEP padding

-> PublicKey
public key

-> ByteString
plaintext

-> Either Error ByteString

Encrypt plaintext using public key an a predefined OAEP seed.

See algorithm 8.11 in "Handbook of Applied Cryptography" by Alfred J. Menezes et al.

decrypt

Arguments

:: HashAlgorithm hash

=> OAEPParams hash ByteString ByteString
OAEP padding parameters

-> PrivateKey
private key

-> ByteString
ciphertext

-> Maybe ByteString

Decrypt ciphertext using private key.

See algorithm 8.12 in "Handbook of Applied Cryptography" by Alfred J. Menezes et al.

sign

Arguments

:: (MonadRandom m, HashAlgorithm hash)

=> PrivateKey
private key

-> hash
hash function

-> ByteString
message to sign

-> m (Either Error Signature)

Sign message using hash algorithm and private key.

See https://en.wikipedia.org/wiki/Rabin_signature_algorithm.

signWith

Arguments

:: HashAlgorithm hash

=> ByteString
padding

-> PrivateKey
private key

-> hash
hash function

-> ByteString
message to sign

-> Either Error Signature

Sign message using padding, hash algorithm and private key.

See https://en.wikipedia.org/wiki/Rabin_signature_algorithm.

verify

Arguments

:: HashAlgorithm hash

=> PublicKey
private key

-> hash
hash function

-> ByteString
message

-> Signature
signature

-> Bool

Verify signature using hash algorithm and public key.

See https://en.wikipedia.org/wiki/Rabin_signature_algorithm.